Bug Fix
When 'get my location' isn't pressed, when toggled off and then on again, map only fits to bounds on the initial toggle, and does not do it at an interval of 1 second. This function is now called on the toggle. The toggle and the button in effect now do the same thing.

As per the previous commit, the toggle also now calls the function to get location, activating the browsers Geolocate.